# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: INTELTEC Production Control API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /producao:
    get:
      operationId: listProducao
      tags: [producao]
      summary: List production orders with filters
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: clienteId
          in: query
          schema:
            type: integer
        - name: dataInicio
          in: query
          schema:
            type: string
            format: date
        - name: dataFim
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of production orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProducaoWithCliente"
    post:
      operationId: createProducao
      tags: [producao]
      summary: Create a new production order (Recebimento)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProducaoBody"
      responses:
        "201":
          description: Production order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProducaoWithCliente"
  /producao/{id}:
    get:
      operationId: getProducao
      tags: [producao]
      summary: Get a production order with all items
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Production order with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProducaoDetail"
    put:
      operationId: updateProducao
      tags: [producao]
      summary: Update production order status or info
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProducaoBody"
      responses:
        "200":
          description: Production order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProducaoWithCliente"
  /producao/{id}/concluir-processamento:
    post:
      operationId: concluirProcessamento
      tags: [producao]
      summary: Conclude processing stage and advance order status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProducaoWithCliente"
  /producao/{id}/cancelar:
    post:
      operationId: cancelarProducao
      tags: [producao]
      summary: Cancel a production order
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Production order cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProducaoWithCliente"
  /producao/{id}/items:
    get:
      operationId: listProducaoItems
      tags: [producao-items]
      summary: List items of a production order
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProducaoItemWithProduto"
    post:
      operationId: addProducaoItem
      tags: [producao-items]
      summary: Add an item to a production order (Processamento)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProducaoItemBody"
      responses:
        "201":
          description: Item added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProducaoItemWithProduto"
  /producao/{id}/items/{itemId}:
    put:
      operationId: updateProducaoItem
      tags: [producao-items]
      summary: Update a production item
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProducaoItemBody"
      responses:
        "200":
          description: Item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProducaoItemWithProduto"
    delete:
      operationId: deleteProducaoItem
      tags: [producao-items]
      summary: Remove an item from a production order
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /impressao/items:
    get:
      operationId: listImpressaoItems
      tags: [producao-items]
      summary: List items pending printing (produto.impresso=true, not yet printed)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Items pending printing
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProducaoItemWithProduto"
  /impressao/marcar:
    post:
      operationId: marcarImpresso
      tags: [producao-items]
      summary: Mark items as printed
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarcarItemsBody"
      responses:
        "200":
          description: Items marked as printed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /envelopamento/items:
    get:
      operationId: listEnvelopamentoItems
      tags: [producao-items]
      summary: List items pending enveloping
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Items pending enveloping
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProducaoItemWithProduto"
  /envelopamento/marcar:
    post:
      operationId: marcarEnvelopado
      tags: [producao-items]
      summary: Mark items as enveloped
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarcarItemsBody"
      responses:
        "200":
          description: Items marked as enveloped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /embalagem/items:
    get:
      operationId: listEmbalagemItems
      tags: [producao-items]
      summary: List items ready for packing
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Items ready for packing
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProducaoItemWithProduto"
  /embalagem/marcar:
    post:
      operationId: marcarEmbalado
      tags: [producao-items]
      summary: Mark items as packed
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarcarItemsBody"
      responses:
        "200":
          description: Items marked as packed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /despacho/items:
    get:
      operationId: listDespachoItems
      tags: [producao-items]
      summary: List items ready for dispatch (packed but not dispatched)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Items ready for dispatch
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProducaoItemWithProduto"
  /despacho/marcar:
    post:
      operationId: marcarDespachado
      tags: [producao-items]
      summary: Mark items as dispatched
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarcarItemsBody"
      responses:
        "200":
          description: Items marked as dispatched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Produto:
      type: object
      properties:
        id:
          type: integer
        descricao:
          type: string
        impresso:
          type: boolean
        envelopado:
          type: boolean
        ativo:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - descricao
        - impresso
        - envelopado
        - ativo
        - createdAt
...
    ProducaoStatus:
      type: string
      enum:
        - RECEBIDA
        - EM_PROCESSAMENTO
        - PROCESSADA
        - EM_PRODUCAO
        - EMBALADA
        - FINALIZADA
        - CANCELADA
    ProducaoWithCliente:
      type: object
      properties:
        id:
          type: integer
        clienteId:
          type: integer
        dataRecebimento:
          type: string
          format: date
        observacoes:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/ProducaoStatus"
        cliente:
          $ref: "#/components/schemas/Cliente"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - clienteId
        - dataRecebimento
        - status
        - cliente
        - createdAt
        - updatedAt
    ProducaoDetail:
      type: object
      properties:
        id:
          type: integer
        clienteId:
          type: integer
        dataRecebimento:
          type: string
          format: date
        observacoes:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/ProducaoStatus"
        cliente:
          $ref: "#/components/schemas/Cliente"
        items:
          type: array
          items:
            $ref: "#/components/schemas/ProducaoItemWithProduto"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - clienteId
        - dataRecebimento
        - status
        - cliente
        - items
        - createdAt
        - updatedAt
    CreateProducaoBody:
      type: object
      properties:
        clienteId:
          type: integer
        dataRecebimento:
          type: string
          format: date
        observacoes:
          type: string
      required:
        - clienteId
        - dataRecebimento
    UpdateProducaoBody:
      type: object
      properties:
        clienteId:
          type: integer
        dataRecebimento:
          type: string
          format: date
        observacoes:
          type: string
        status:
          $ref: "#/components/schemas/ProducaoStatus"
    ProducaoItemWithProduto:
      type: object
      properties:
        id:
          type: integer
        producaoId:
          type: integer
        produtoId:
          type: integer
        itemNumero:
          type: integer
        quantidade:
          type: integer
        multiplicador:
          type: number
        impresso:
          type: boolean
        envelopado:
          type: boolean
        embalado:
          type: boolean
        despachado:
          type: boolean
        dataUltimoStatus:
          type: string
          format: date-time
          nullable: true
        produto:
          $ref: "#/components/schemas/Produto"
        producao:
          $ref: "#/components/schemas/ProducaoWithCliente"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - producaoId
        - produtoId
        - itemNumero
        - quantidade
        - multiplicador
        - impresso
        - envelopado
        - embalado
        - despachado
        - produto
        - producao
        - createdAt
    CreateProducaoItemBody:
      type: object
      properties:
        produtoId:
          type: integer
        quantidade:
          type: integer
        multiplicador:
          type: number
      required:
        - produtoId
        - quantidade
    UpdateProducaoItemBody:
      type: object
      properties:
        produtoId:
          type: integer
        quantidade:
          type: integer
        multiplicador:
          type: number
    MarcarItemsBody:
      type: object
      properties:
        itemIds:
          type: array
          items:
            type: integer
      required:
        - itemIds
...